chore: move MockAffiliateRegistry to cf_traits
