+++ state-chain/traits/src/mocks.rs
@@ -5,6 +5,7 @@ use frame_support::{storage, StorageHasher, Twox64Concat};
 
 pub mod account_role_registry;
 pub mod address_converter;
+pub mod affiliate_registry;
 pub mod api_call;
 pub mod asset_converter;
 pub mod asset_withholding;
